refactor: rename fluid factory file
